+++ Changelog.md
@@ -5,6 +5,7 @@
 ### Added
   - Support generics in array size instance
   - Generics definition now supports an optional description
+  - Add new command-line argument "--check" to only run compilation (no generator called)
 
 ### Fixed
   - Override: for register array, indexed and default override are now merged (previously default was fully ignored even without indexed override)